requirements.txt
fastapi
uvicorn
supabase
python-dotenv
fpdf
python-multipart